refactor: 회원 등급 삭제 시 오류 메시지 처리
