--- src/main/java/shop/nuribooks/view/member/dto/response/MemberUpdateResponse.java
@@ -1,5 +1,0 @@
-package shop.nuribooks.view.member.dto.response;
-
-public class MemberUpdateResponse {
-
-}
