+++ src/main/java/shop/nuribooks/view/member/member/dto/response/MemberDetailsResponse.java
@@ -1,4 +1,4 @@
-package shop.nuribooks.view.member.dto.response;
+package shop.nuribooks.view.member.member.dto.response;
 
 import static com.fasterxml.jackson.annotation.JsonFormat.Shape.*;
