+++ src/main/java/shop/nuribooks/view/member/member/service/MemberServiceImpl.java
@@ -1,4 +1,4 @@
-package shop.nuribooks.view.member.service;
+package shop.nuribooks.view.member.member.service;
 
 import java.math.BigDecimal;
 import java.util.Objects;
@@ -12,12 +12,12 @@ import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import shop.nuribooks.view.common.dto.ResponseMessage;
-import shop.nuribooks.view.member.dto.request.MemberPasswordUpdateRequest;
-import shop.nuribooks.view.member.dto.response.MemberDetailsResponse;
-import shop.nuribooks.view.member.feign.MemberServiceClient;
+import shop.nuribooks.view.member.member.dto.request.MemberPasswordUpdateRequest;
+import shop.nuribooks.view.member.member.dto.response.MemberDetailsResponse;
+import shop.nuribooks.view.member.member.feign.MemberServiceClient;
 import shop.nuribooks.view.common.util.ExceptionUtil;
-import shop.nuribooks.view.member.dto.request.MemberRegisterRequest;
-import shop.nuribooks.view.member.dto.response.MemberRegisterResponse;
+import shop.nuribooks.view.member.member.dto.request.MemberRegisterRequest;
+import shop.nuribooks.view.member.member.dto.response.MemberRegisterResponse;
 
 /**
  * 회원 서비스 구현체
