+++ src/main/resources/templates/admin/grade/grade-register.html
@@ -27,4 +27,9 @@
             </div>
         </div>
     </div>
+
+    <!-- 회원등급 목록으로 돌아가기 링크 -->
+    <div class="mt-3 text-center">
+        <a href="/admin/grade" class="btn btn-secondary">회원등급 목록으로 돌아가기</a>
+    </div>
 </div>
